CASMCMS-7388: Transitioning to Github and Artifactory
As part of the transition to Github CMS is pushing its artifacts to its
own artifactory. This commit references a base Docker image from this
new artifactory location algol60. It also changes URLs to point to
Github rather than Stash.

A newer version of the install_cms_meta_tools.sh script was used.